+++ src/driftbuster/formats/dockerfile/__init__.py
@@ -1,0 +1,6 @@
+"""Dockerfile format plugin package."""
+
+from .plugin import DockerfilePlugin
+
+__all__ = ["DockerfilePlugin"]
+