+++ src/driftbuster/formats/yaml/__init__.py
@@ -1,0 +1,6 @@
+"""YAML format plugin package."""
+
+from .plugin import YamlPlugin
+
+__all__ = ["YamlPlugin"]
+